LayoutTests/imported/w3c/web-platform-tests/cookies/samesite/about-blank-subresource.https-expected.txt


FAIL SameSite cookies on subresource of top-level about:blank window assert_equals: `samesite_unspecified=0.9309266340228124` in request to `https://web-platform.test:9443`. expected (string) "0.9309266340228124" but got (undefined) undefined
